requirements.txt
streamlit==1.46.0

tf-nightly

opencv-python

numpy

pillow

matplotlib

scikit-learn

sqlalchemy

gdown  # <-- THIS fixes the "ModuleNotFoundError: No module named 'gdown'" seen in your latest error
